requirements.txt
# python 3.13
msgpack==1.1.0
pyyaml==6.0.2
Pillow==11.2.1
flask==3.1.1
flask-jwt-extended==4.7.1